# Cherry-pick #6343 to 6.2: Optimize filebeat state ID generation