Cherry-pick of PR #6343 to 6.2 branch. Original message: 

use append operations (StringBuilder pattern) instead of fmt.Sprintf to
reduce allocation overhead when generating the id.

The benchmark generates 2500 ids per run. As the ID is a string, one
at least one allocation per ID is required. Results show the change cuts
allocations by 3.

```
benchmark                  old ns/op     new ns/op     delta
BenchmarkStateString-4     633983        274193        -56.75%
BenchmarkStateString-4     634008        275098        -56.61%
BenchmarkStateString-4     627028        273854        -56.33%
BenchmarkStateString-4     629635        275542        -56.24%
BenchmarkStateString-4     625108        284221        -54.53%
BenchmarkStateString-4     632816        272703        -56.91%
BenchmarkStateString-4     633756        275215        -56.57%
BenchmarkStateString-4     671575        273220        -59.32%
BenchmarkStateString-4     663687        272034        -59.01%
BenchmarkStateString-4     629798        279793        -55.57%

benchmark                  old allocs     new allocs     delta
BenchmarkStateString-4     7400           2500           -66.22%
BenchmarkStateString-4     7400           2500           -66.22%
BenchmarkStateString-4     7400           2500           -66.22%
BenchmarkStateString-4     7400           2500           -66.22%
BenchmarkStateString-4     7400           2500           -66.22%
BenchmarkStateString-4     7400           2500           -66.22%
BenchmarkStateString-4     7400           2500           -66.22%
BenchmarkStateString-4     7400           2500           -66.22%
BenchmarkStateString-4     7400           2500           -66.22%
BenchmarkStateString-4     7400           2500           -66.22%

benchmark                  old bytes     new bytes     delta
BenchmarkStateString-4     157619        118388        -24.89%
BenchmarkStateString-4     157623        118387        -24.89%
BenchmarkStateString-4     157630        118387        -24.90%
BenchmarkStateString-4     157631        118387        -24.90%
BenchmarkStateString-4     157631        118387        -24.90%
BenchmarkStateString-4     157631        118387        -24.90%
BenchmarkStateString-4     157631        118387        -24.90%
BenchmarkStateString-4     157631        118387        -24.90%
BenchmarkStateString-4     157631        118387        -24.90%
BenchmarkStateString-4     157630        118387        -24.90%
```